refactor(portal): change fille structure in layout component
affects: @fremtind/portal
