+++ portal/src/components/Layout/ComponentLayout.tsx
@@ -1,6 +1,6 @@
 import React, { ReactNode } from "react";
 import { ReactIcon, SassIcon } from "../icons";
-import { Layout } from "./layout-react/src";
+import { Layout } from ".";
 import { CodeLink } from "./CodeLink";
 import "../layout/Layout.scss";
