+++ portal/src/components/index.ts
@@ -1,5 +1,5 @@
 import { Header } from "./Header";
-import { Layout } from "./Layout/layout-react/src";
+import { Layout } from "./Layout";
 import { ComponentLayout } from "./Layout/ComponentLayout";
 import { Menu } from "./Menu/Menu";
